Add validation config, update imports
